Celery Executor load time optimizations
Import optimizations to decrease the amount of time it takes to
load/import the celery executor.
Move some expensive typing related imports to be under TYPE_CHECKING.

Refactor expensive classes (other than the CeleryExecutor) out of the
celery_executor.py module into a utils module so they are only
loaded at runtime.

Make the Celery app still accessible by leveraging PEP562

Also move some imports to runtime imports, closer to their usage.
